+++ inst/application/uitpInput.R
@@ -108,14 +108,14 @@ uitpInput <-
         popify(textInput("minParentAge", label = "Minimum Parent Age (years)",
                   value = "0.0"), NULL,
                paste("Parents must be at least as old as the minimum parent",
-                     "age on the birth date of an offspring. If not, the",
-                     "filed will not be accepted and a file named",
+                     "age at the birthdate of an offspring. If not, the",
+                     "file will not be accepted and a file named",
                      "<b>lowParentAge.csv</b> containing a list of parents",
                      "below the minimum age will",
-                     "be written to the users home directory. Animals",
+                     "be written to the user's home directory. Animals",
                      "without birth dates are not affected by this rule.")),
 
-        actionButton("getData", "Read and Test Pedigree"),
+        actionButton("getData", "Read and Check Pedigree"),
         checkboxInput("debugger", label = "Debug on", value = FALSE)#,
         ),
       # Main Panel